Moved model service unit tests to api-common
